Wallet recover CLI Manager migration
