+++ validator/accounts/cli_manager.go
@@ -25,7 +25,7 @@ func NewCLIManager(opts ...Option) (*AccountsCLIManager, error) {
 }
 
 // AccountsCLIManager defines a struct capable of performing various validator
-// wallet account operations via the command line.
+// wallet & account operations via the command line.
 type AccountsCLIManager struct {
 	wallet               *wallet.Wallet
 	keymanager           keymanager.IKeymanager
@@ -48,6 +48,11 @@ type AccountsCLIManager struct {
 	filteredPubKeys      []bls.PublicKey
 	rawPubKeys           [][]byte
 	formattedPubKeys     []string
+	walletDir            string
+	walletPassword       string
+	mnemonic             string
+	numAccounts          int
+	mnemonic25thWord     string
 }
 
 func (acm *AccountsCLIManager) prepareBeaconClients(ctx context.Context) (*ethpb.BeaconNodeValidatorClient, *ethpb.NodeClient, error) {
